Replace package{} with ensure_packages()
Use `ensure_packages()` from stdlib to specify the package resource. This
will prevent it from conflicting with a real PAM module. It would have been
alright if we could pin a small dependency like `libpam-motd`, but since
that's not possible.

I dislike the essence of `ensure_packages()` but it should do the job here.