Add purge_directory param
Allow the unamanged contents of `/etc/update-motd.d` to be purged. This will
be of use if you define scripted content using a (yet to come) defined type
and want it to be removed when you omit the resources in Puppet. We might
need to be careful not to wipe out the files that Ubuntu places there by
default, unless you really want to.