Update the module to pass tests under puppet 4.
These are all in a single PR as you need to update the rspec-puppet
gem and tests in lockstep to enable puppet 4 support.

Puppet changed its behaviour in regards to converting bare numbers in
3 to strings, which puppet 4 does not do. The order parameter was
impacted by this so we now force it in to a string so it can be
used in either puppet version. Note that the specs do not change.

The raise syntax was changed upstream to the new form we're using below.